tests/extraction-benchmark/requirements.txt
google-generativeai>=0.8.0
Pillow>=10.0.0
numpy>=1.24.0
python-Levenshtein>=0.21.0
scipy>=1.11.0
aiohttp>=3.9.0
pdf2image>=1.16.0
pydantic>=2.0.0
